apps/backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name", email = "you@example.com"},
]
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "sqlalchemy",
    "psycopg2-binary",
    "python-dotenv",
    "alembic",
    "yfinance",
    "sqlmodel",
    "ib_async",
    "openpyxl>=3.1.5",
    "scipy>=1.15.3",
    "numpy>=2.2.6",
    "python-dateutil",
    "opentelemetry-distro",
    "opentelemetry-exporter-otlp",
    "opentelemetry-instrumentation-fastapi",
    "opentelemetry-instrumentation-sqlalchemy",
    "opentelemetry-instrumentation-logging",
]
requires-python = ">=3.10"

[tool.setuptools.packages.find]
include = ["app*"]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.uv]
dev-dependencies = [
    "typing_extensions",
    "mako",
]
